// Extra_Activities/dict/api/extraactivities/types.ts
export type PostExtraactivitiesCreateResponse = PostExtraactivitiesCreate201Response | PostExtraactivitiesCreate401Response | PostExtraactivitiesCreate404Response | PostExtraactivitiesCreate422Response

export interface PostExtraactivitiesCreate201Response {
	status: 201
	body: Api.ExtraactivitiesDto
	headers?: never
}

export interface PostExtraactivitiesCreate401Response {
	status: 401
	body: Api.MessageDto
	headers?: never
}

export interface PostExtraactivitiesCreate404Response {
	status: 404
	body: Api.MessageDto
	headers?: never
}

export interface PostExtraactivitiesCreate422Response {
	status: 422
	body: Api.MessageDto
	headers?: never
}

export type DeleteExtraactivitiesDeleteResponse = DeleteExtraactivitiesDelete200Response | DeleteExtraactivitiesDelete401Response | DeleteExtraactivitiesDelete404Response | DeleteExtraactivitiesDelete422Response

export interface DeleteExtraactivitiesDelete200Response {
	status: 200
	body: Api.MessageDto
	headers?: never
}

export interface DeleteExtraactivitiesDelete401Response {
	status: 401
	body: Api.MessageDto
	headers?: never
}

export interface DeleteExtraactivitiesDelete404Response {
	status: 404
	body: Api.MessageDto
	headers?: never
}

export interface DeleteExtraactivitiesDelete422Response {
	status: 422
	body: Api.MessageDto
	headers?: never
}

export type GetExtraactivitiesGetResponse = GetExtraactivitiesGet200Response | GetExtraactivitiesGet401Response | GetExtraactivitiesGet404Response | GetExtraactivitiesGet422Response

export interface GetExtraactivitiesGet200Response {
	status: 200
	body: Api.ExtraactivitiesDto
	headers?: never
}

export interface GetExtraactivitiesGet401Response {
	status: 401
	body: Api.MessageDto
	headers?: never
}

export interface GetExtraactivitiesGet404Response {
	status: 404
	body: Api.MessageDto
	headers?: never
}

export interface GetExtraactivitiesGet422Response {
	status: 422
	body: Api.MessageDto
	headers?: never
}

export type GetExtraactivitiesGetAllResponse = GetExtraactivitiesGetAll200Response | GetExtraactivitiesGetAll401Response | GetExtraactivitiesGetAll404Response

export interface GetExtraactivitiesGetAll200Response {
	status: 200
	body: Api.ExtraactivitiesPagedResultDto
	headers?: never
}

export interface GetExtraactivitiesGetAll401Response {
	status: 401
	body: Api.MessageDto
	headers?: never
}

export interface GetExtraactivitiesGetAll404Response {
	status: 404
	body: Api.MessageDto
	headers?: never
}

export type PutExtraactivitiesUpdateResponse = PutExtraactivitiesUpdate200Response | PutExtraactivitiesUpdate401Response | PutExtraactivitiesUpdate404Response | PutExtraactivitiesUpdate422Response

export interface PutExtraactivitiesUpdate200Response {
	status: 200
	body: Api.ExtraactivitiesDto
	headers?: never
}

export interface PutExtraactivitiesUpdate401Response {
	status: 401
	body: Api.MessageDto
	headers?: never
}

export interface PutExtraactivitiesUpdate404Response {
	status: 404
	body: Api.MessageDto
	headers?: never
}

export interface PutExtraactivitiesUpdate422Response {
	status: 422
	body: Api.MessageDto
	headers?: never
}